Angle base class
