+++ Arm/project.mk
@@ -1,5 +1,5 @@
-TESTS += test/wrist_joint_test.cpp
+#TESTS += test/wrist_joint_test.cpp
 TESTS += test/rover_arm_system_test.cpp
-TESTS += test/joint_test.cpp
-TESTS += test/finger_test.cpp
-TESTS += test/hand_test.cpp
+#TESTS += test/arm_joint_test.cpp
+#TESTS += test/finger_test.cpp
+#TESTS += test/hand_test.cpp
